Refactored components to simplify directories
