--- components/Home/OscarTitle/LineTitle.tsx
@@ -1,8 +1,0 @@
-import type { ReactElement } from 'react'
-import style from '../../../styles/home/oscar-title/LineTitle.module.scss'
-
-export default function LineTitle (): ReactElement {
-  return (
-    <div className={style.container}></div>
-  )
-}
