--- components/Home/TopMenu/Divider.tsx
@@ -1,8 +1,0 @@
-import type { ReactElement } from 'react'
-import style from '../../../styles/home/top-menu/Divider.module.scss'
-
-export default function Divider (): ReactElement {
-  return (
-    <div className={style.divider}></div>
-  )
-}
