--- components/Home/TopMenu/ShadowGradient.tsx
@@ -1,8 +1,0 @@
-import type { ReactElement } from 'react'
-import style from '../../../styles/home/top-menu/ShadowGradient.module.scss'
-
-export default function ShadowGradient (): ReactElement {
-  return (
-    <div className={style.shadow}></div>
-  )
-}
