+++ components/TopMenu/TopMenu.tsx
@@ -1,11 +1,9 @@
 import type { ReactElement } from 'react'
-import style from '../../../styles/home/top-menu/TopMenu.module.scss'
+import style from '../../styles/top-menu/TopMenu.module.scss'
 import SocialMedia from './SocialMedia'
 
-import logo from '../../../public/logo.webp'
+import logo from '../../public/logo.webp'
 import Image from 'next/image'
-import Divider from './Divider'
-import ShadowGradient from './ShadowGradient'
 
 export default function TopMenu (): ReactElement {
   return (
@@ -22,8 +20,8 @@ export default function TopMenu (): ReactElement {
         />
         <SocialMedia />
       </div>
-      <Divider />
-      <ShadowGradient />
+      <div className={style.divider}></div>
+      <div className={style.shadow}></div>
     </div>
   )
 }
