+++ pages/index.tsx
@@ -1,6 +1,6 @@
 import type { ReactElement } from 'react'
 import HomePage from '../components/Home/HomePage'
-import style from '../styles/oscar.module.scss'
+import style from '../styles/home/home.module.scss'
 
 export default function Home (): ReactElement {
   return (
